slack->mutl-accountある程度終了\n\nあとはfrontendやテストなどかしら.構成見直しも含めて.
