+++ memo.md
@@ -1,2 +1,5 @@
 TODO: database migration
-TODO: Signup時のメールの修正、会費支払い確認メール送信 
+
+
+TODO: Signup時のメールの修正(入力欄をなくすor最初から入力させとく)、会費支払い確認メール送信 
+TODO: ログインしたアカウントがDBになかった時の処理
